Fix column table tests
